generated_projects/gce_discord_bot/gce_discord_bot/requirements.txt
discord.py==2.0.0
google-cloud-sdk==0.13.0